Remove lclIdxType from my block distribution experiments
As we discussed at the code review a few weeks back, my
implementation was not correct enough with its use of
locIdxType and glbIdxType (in particular, it was assumed
that values could be converted back and forth between
them willy-nilly), and that making it correct would
complicate the implementation more than it was worth at
this stage of the distribution studies.  For this
reason, replaced the global and local index types with
a single index type idxType.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14430 3a8e244f-b0f2-452b-bcba-4c88e055c3ca