enabled ~ on bool
This is listed in the spec, but should we really support it?  This
operator is only supported over int(32), int(64), uint(32), and
uint(64) in C#.  Since we coerce bool to these types, it would likely
still work.  One problem with supporting ~ on bool is that ~true
remains true.  Isn't this a problem in the current implementation?
I'm checking in because Brad's futures expect this behavior, but I'd
like to know the reason why?  My leaning is that ~ should be supported
over bool, but it should be defined to return false if the operand is
true, and vice versa.

In another note, I'm concerned by the bool types of different widths
for two reasons.  First, the circularity in the implicit conversions
is unnerving.  It is unclear to me why it seems natural that there is
no error in the test

  test/types/scalar/bradc/overloadbools.chpl

but there is in

  test/types/scalar/bradc/overloadbools2.chpl

I understand why this is currently happening, but I'm not sure it
seems appropriate.

Second, the behavior of bitwise operators on bools that use multiple
bits seems potentially bizarre especially as these values can then be
converted to integers.

Removed two futures.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14435 3a8e244f-b0f2-452b-bcba-4c88e055c3ca