Added error check to require parentheses on 'this' and 'these' methods and added parentheses to the 'these' method in the standard Vector class.  I realize that Andy probably leaves this off because, given a class instance c, to iterate over it one writes
  for e in c do

but to access via 'this' one writes

  c()

That makes me think that we should leave the parentheses off of the
'these' method except that for leader/follower iterators, we will need
parentheses.  So hopefully this new error message will catch this case.

Also, removed an unnecessary AST collection in the checkParsed pass.

Removed a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14436 3a8e244f-b0f2-452b-bcba-4c88e055c3ca