Minor simplifications of coercion wrapper code
Also factored out the setting of wrapper->isCompilerTemp to true so
that it is done for all wrappers except the default wrapper for a
constructor; this makes sense because the default wrapper for a
constructor is more of a specialized constructor.  Setting the
isCompilerTemp flag for wrappers improves error message reporting
since a call site of the wrapper is used for function resolution
errors.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14448 3a8e244f-b0f2-452b-bcba-4c88e055c3ca