Make range slicing work near type boundaries
The code to snap the low and high range bounds in place after slicing it with
another range would overflow when working near min(eltType). The overflow was
happening in the code: var d = abs(alignment - low) % s; where alignment and
low are both of generic type eltType, and s is a positive int cast to eltType.

Since s is positive, this overflow can be avoided by pushing the modulus in
before the subtraction using a few tricks.

For negative a, the a % b operation doesn't return a number that is congruent
to a (mod b), so I wrote a mod(a,b) function that returns the value in [0, b-1]
that is congruent to a (mod b). This could maybe go into the Math module.

First, note that if a > b then abs(a - b) is a - b, otherwise it is b - a. So,
replace the abs() call with this if then else expression.

Next, if a1 is congruent to a2 (mod s) and b1 is congruent to b2 (mod s) then
a1 - b1 is congruent to a2 - b2 (mod s). Use the mod function defined above to
create suitable a2 and b2, and subtract these (mod s) instead.

Finally, with the uint types, the a2 - b2 subtraction could still result in an
overflow. This can be avoided by seeing that they are both in [0, s-1] and s
is a positive int, so a2 and b2 can safely be cast to int before doing the
subtraction.

Remove my .future about this. Also, fix the .good for my Cyclic distribution
test. It was expecting to be one element smaller than it should have been
because of these overflow problems.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14450 3a8e244f-b0f2-452b-bcba-4c88e055c3ca