Added support for printing list of executing and pending tasks.
Programs that are executed with the -t or --taskreport flag print this report.
When used in conjunction with -b the task report will also be printed when
deadlock occurs.

I also had to modify a number of tests that check the output message when a
chapel program is invoked with --help (since it now includes a line about the
new taskreport feature).

This change also introduces the ChapelRuntime standard module.  ChapelRuntime
is a module of functions that are useful to call from within the Chapel
runtime.  It's there for the convienence of Chapel's developers (so we aren't
always stuck writing runtime functionality in C).

I also added a number of tests to check task report output (under the parallel
directory).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14467 3a8e244f-b0f2-452b-bcba-4c88e055c3ca