Error message future: promotion/iterator reported where none exists
This test seems like it's generating a bad error message.  It's
complaining about promotion/iterator, yet I don't think the
test uses either.  I spent a fair amount of time struggling with
this in my distribution work before realizing what was causing
the problem.  See the comments in the code and .future for more
details.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14468 3a8e244f-b0f2-452b-bcba-4c88e055c3ca