Chained destructors to their parents destructors
Now, when a destructor is called, the class hierarchy is searched for a parent
class with a destructor.  If one is found (such as the one for the object
class), it is called.

In order to avoid infinite recursion, a new primitive called PRIMITIVE_DELETE
was introduced.  This primitive is used when a delete statement is encountered,
and during normalization, it is changed into a call to ~chpl_destroy.  Dynamic
dispatch is disabled for destructors during function resolution.

The compiler no longer inserts calls to free memory; instead, the destructor
for the object class explicitly calls chpl_free().  Since all destructors
inherit from the object class, the destructor for the object class is eventually
called.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14471 3a8e244f-b0f2-452b-bcba-4c88e055c3ca