Added labels to classes in BradsBlock1DPar classes
As we discussed during the last code review, I added some
high-level labels in the form of comments to my six classes
so that the different parts of the classes would be clearer.
In particular:

  // GENERICS:     the types/params that make the class generic
  // STATE:        the local state of the class
  // LINKAGE:      linkage to other classes that make up the distribution
                     DOWN = link from global class to local
                     UP   = link from local class to global
                     LEFT = link from dom->dist or from arr->dom
  // CONSTRUCTORS: constructors/initialize calls
  // INTERFACE:    different aspects of the interface
                     INTERNAL = specific to my block distribution/use of it
                     <other>  = interface for the named distribution class

This should help us read these classes, and should also help me
try to massage them into an official distribution, with any luck.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14473 3a8e244f-b0f2-452b-bcba-4c88e055c3ca