Add Dimensional distribution
Commit what I have so far for a dimension by dimension distribution. It works
when each dimension on each locale can be described with a single range.
For example, a cyclic dimension works and a blocked dimension works, but a
block-cyclic dimension does not, as it can't be represented with a single
range.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14483 3a8e244f-b0f2-452b-bcba-4c88e055c3ca