Fixing subdirectories under lib/
Recently noticed that our libraries were not being pushed
as far down under lib/ as we'd intended, probably as a
result of my last Makefile refactoring.  In particular,
the platform and compiler were not being used to create
the lib/ subdirectory.  This fixes that issue and factors
the lib/ path into the Makefile.base directory to avoid
repetition.

Also updated README.files to more accurately illustrate
the lib/ hierarchy.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14487 3a8e244f-b0f2-452b-bcba-4c88e055c3ca