Moved modules/standard/ChapelRuntime.chpl to runtime/src/chapel_code.chpl so that we can compile it when building the runtime instead of with every Chapel program.  Added a list of standard modules to include in this stand-alone compile.
Added compiler option --runtime which builds a .h and .c file instead
of an executable ensuring that all functions defined with a pragma
"export" are available in the .h file For the purpose of resolution,
these are considered possible entry points to running Chapel code.
These stand-alone files must be linked against the runtime.  Used this
option to build chapel_code.chpl above.  This required the following
changes to the compiler:

 - Added openRuntimeFile routine to files.cpp in order to generate a
   file in the same place that the compiler is invoked (rather than in
   the --savec or temporary directory).

 - Disable the creation of initModuleGuards and heapAllocateGlobals in
   the generated code when --runtime is used.

 - Disable the insertion of a 'use' of ChapelStandard in theProgram
   when --runtime is used.

 - Make resolve and pruning aware of pragma "export" when --runtime is
   used.

 - Made code generation generate all code into the stand-alone .h and
   .c files rather than into files in the --savec or temporary
   directory and disabled building of the binary when --runtime is
   used.

 - Disable the generation of maxThreads and all other config variables
   in ChapelBase and clear the ChapelBase initialization function so
   that it doesn't do anything when --runtime is used.

 - Changed main into an initialization function for the stand-alone .c
   file when --runtime is used.  This should be called before any of
   its code is used.  Simplified main so that it doesn't use endCount
   to ensure that all threads complete.  For chapel_code, I added a
   call to chpl_init_chapel_code in the main routine in the runtime.

Added (temporarily) runtime/src/chapel_code.c and
runtime/include/chapel_code.h.  These eventually should not be part of
the system but built using --runtime --no-checks and chpl, but I
didn't want to deal with the 'make depend' issues that Brad foresees.
Instead, the runtime makefiles have a special rule for this C file
only because the generated code requires different warning flags.

This should improve compilation speed again since we now compile the
task-list hash table when building the runtime instead of when
compiling the generated code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14488 3a8e244f-b0f2-452b-bcba-4c88e055c3ca