Revamped implementation of pragmas.  These are now interfaced in the compiler with an enumeration and their implementation uses a bit vector instead of a vector of strings.  To add a pragma (documented in pragmas.h), add an enumeration entry to PragmaTag and then define it in a call to definePragma in the function initPragmas(); this latter step adds it to the map from strings to integers so that the pragma can be added to a symbol in Chapel code.  Even if the pragma is not going to be used in Chapel code, it should still be defined so that it can be understood when debugging.
Because debugging pragmas is no longer as clear without the vector of
strings, I added a GDB command called pragmas that takes an ast and
prints out the pragmas associated with it.  For example, "pragmas fn"
in GDB will print out the list of pragmas that apply to the symbol fn.

Changed bools Symbol::isExtern and FnSymbol::noParens into pragmas.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14501 3a8e244f-b0f2-452b-bcba-4c88e055c3ca