Fixing GASNet segfault
Due to a recent refactoring, several GASNet tests seg faulted due
to dereferencing a NULL task list to get filename/lineno information.
I guarded this dereference and set to <unknown>:0 for now in the case
of a NULL pointer (which we believe happens only in the case of an on
clause).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14503 3a8e244f-b0f2-452b-bcba-4c88e055c3ca