Specialized CHPL_MAKE_PLATFORM Makefile variable
Previously, any given Makefile would either be in host or
target mode, and CHPL_MAKE_PLATFORM would be set appropriately.
With this checkin, I added variables CHPL_MAKE_HOST_PLATFORM
and CHPL_MAKE_TARGET_PLATFORM so that we can distinguish
between host and target when necessary.  For example, when
building the runtime, the code is being created for the target,
but we need to use the host to find the location of the Chapel
compiler.

Unfortunately, not all references to CHPL_MAKE_PLATFORM can
be removed from the current Makefiles because of the fact that
we want our compiler settings to be usable for either a source
or a target.  Put another way, I don't think we have a clean
way to create CCFLAGS_HOST and CCFLAGS_TARGET, so we just
create CCFLAGS based on CHPL_MAKE_PLATFORM, which is set based
on what the current Makefile is aimed at creating.  That said,
I've tried to remove all references to CHPL_MAKE_PLATFORM that
don't need to be flexible in this way.

Also, removed a duplicate definition of GASNET_INSTALL_DIR as
GASNET_INSTALL in the Gasnet Makefiles.  And split
PLAT_COMP_SUBDIR into PLAT_COMPILER_SUBDIR and PLAT_RUNTIME_SUBDIR
in order to minimize the number of references to CHPL_MAKE_PLATFORM
and make things a little more explicit.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14515 3a8e244f-b0f2-452b-bcba-4c88e055c3ca