Removed one bool field and changed two other bool fields into pragmas.
Changed FnSymbol::isWrapper into PRAG_WRAPPER used to mark a function
as a compiler-created wrapper.

Changed FnSymbol::visible into PRAG_INVISIBLE_FN used to mark a
function so that it is not a candidate for function resolution.

Removed FnSymbol::global.  This used to mark functions that were
globally visible.  Back in the days of multiple dispatch, any function
with a class argument was marked as global.  It was currently used to
mark iterator class methods, but these are already specially marked,
so this change uses that marker instead.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14516 3a8e244f-b0f2-452b-bcba-4c88e055c3ca