Improving cross-compilation launch capabilities; added xmt-sim platform
As noted in my "launcher" check in log from quite some time
ago, I had made the simplifying assumption that the source
and target platforms used for the launcher and launchee
binaries were the same in order to simplify the changes that
I needed to make for the Makefile.  These changes were somewhat
similar to the ones that I had to make for the chapel_code.c
compilation changes this morning, so I went ahead and tackled
them as well.  In particular, like the chapel_code.c change,
the launcher script is funny in that it wants to be compiled
in host mode while the launchee wants to be compiled in target
mode.  Since any given Makefile only has a single set of CFLAGS,
etc. this wasn't possible in our Makefile architecture in any
simple manner.

I approached this by moving the runtime/Makefile into a helper
Makefile -- runtime/Makefile.help.  The top-level Makefile now
takes any top-level command and invokes Makefile.help twice --
once for the launcher, and once for the launchee.  Makefile.help
puts itself into host mode for the launcher and target mode for
the launchee, and therefore gets all the right settings.  The
only other thing that remained was to make each directory
specialize its TARGETS variable based on whether it was the
launcher or launchee.  As long as I was doing this, I also
specialized the SUBDIRS variable such that any directories that
are not used for launcher/launchee are no longer descended into.
Thus, launch-none should not be compiled for a standard build
anymore.

The only problem that I've been able to find so far is that
doing a "make" from the runtime directory (rather than a "make all")
causes the runtime to be built twice.  I believe this is because
it's matching my % pattern once for the "default" rule and a
second time for the "all" rule that the default rule says should
be built.  Need to think on this one a bit more, but have made the
top-level Makefile use a "make all" in the runtime directory to
avoid it.

As motivation for getting this to work, I worked on the most
familiar cross-compiler machine for me (the xmt), and created
an xmt-sim target platform similar to the x1-sim/x2-sim that
we've talked about doing for David.  Generally speaking, xmt-sim
works like xmt/mta, except that it has a default launch command
of "zebra" -- the xmt simulator.  Put together this launch
command in fairly short order without too much trouble.  Thus,
at present, if CHPL_TARGET_PLATFORM is set to xmt-sim, the
runtime should get built twice -- once for the launcher and the
xmt and when compiling a program, a linux a.out and a zebra
a.out_real file should be created.  Unfortunately, the a.out_real
file doesn't quite work as one would like within zebra because
of the call to sysctl that's done at program startup time to
set the maximum number of threads.  Will special-case this in
a future checkin and work on making zebra work a bit more
seamlessly.

Rejiggered a few of the Makefiles: 

* fixed some missing variables relating to the launcher for the 
  comm-* directories;

* split the runtime/etc/Makefile.include Makefile into two
  Makefiles in order to get two sets of CC/CFLAGS/etc. -- one
  for the launcher, one for the launchee

* since one needs two versions of "ar" and "ranlib" in order
  to compile libraries for the launcher and launchee, specialized
  the Makefile.linux64 file to give an explicit path to the
  gnu-compatible tools when the target platform is xmt-sim.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14519 3a8e244f-b0f2-452b-bcba-4c88e055c3ca