Slight cleanup to previous checkin
Not only was my % command catching the "default" command
twice, it was also causing a "-include ..." error because
the non-existent Makefile being included would try to be
built using the % rule.  Fixed this by only exposing our
normal targets in the top-level Makefile and using a
more specialized % rule to do the recursive make steps.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14520 3a8e244f-b0f2-452b-bcba-4c88e055c3ca