Getting xmt-sim working
Got xmt-sim working -- changed a call of sysctl() that
was used to query the number of processors into the
more MTA-specific mta_get_num_teams() which seems
equally valid and works within the simulator (sysctl()
is not implemented in the simulator and it seemed
nicer to use one code for both cases than to special-case
the simulator).

Also, changed the invocation of zebra to include the -r
flag which says "run this program and exit when you're
done".

At this point, a few extra warnings are generated, but
otherwise, I think this has now successfully wrapped
the mta launch as one would expect.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14522 3a8e244f-b0f2-452b-bcba-4c88e055c3ca