Removed some leading-_ identifiers
I misdiagnosed the MTA/SunOS compilation failures today.
It wasn't due to the use of the STL (at least not
due to lack of support for it), but rather a conflict
on the variable name _N.  Renamed this and other
leading-underscore names in copyPropagation.cpp to
avoid this issue.  My hypothesis is that the new STL
headers that are being included define _N, but that's
just a guess.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14526 3a8e244f-b0f2-452b-bcba-4c88e055c3ca