Fixing nightly tests for building compiler for mta/xmt
Because the runtime now executes the Chapel compiler as part
of the build process, the nightly regression runs that
compile the compiler for the xmt/mta no longer work (the
compiler can't be run effectively on the mta because (a)
it's too slow, and (b) CHPL_HOME is not available to it
by default).  Fixed this by adding a mode to the nightly
script -noruntime which is like -notest in that it kicks
out of the nightly script early -- after building the
compiler.  Since the runtime is built for the MTA/XMT
on our normal test, this doesn't reduce coverage.

This general issue does raise questions about our
general cross-compilation strategy, though.  If one
is building for a cross-compilation environment in
which target platform binaries can't be run on the
source platform, how should we execute the compiler
to build the runtime files?

Developers may also notice that with the removal
of the relative path to chapel_code.h, doing a
"make depend" prior to a build will generate a
bad dependence on files that include chapel_code.h
from a different directory.  The combination of
these two things may cause us to need to re-evaluate
whether we should be recompiling chapel_code.chpl
on every build.  I don't have an alternative that
I'm happy with in mind yet.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14528 3a8e244f-b0f2-452b-bcba-4c88e055c3ca