Fixing make depend step for chapel_code.h from clean build
Once the relative path was removed for #inclusions of
chapel_code.h from directories other than runtime/src,
doing a "make clean depend all" ceased to work because
gcc assumes that a header file will be local to the
source file's directory when it doesn't already exist.

Got around this by having our make depend step create a
dummy chapel_code.h header file if the file doesn't
already exist.  Since the src/ directory is built before
the other directories, this ensures that it will exist
before any dependences are calculated for other
directories.  This is a little bit hacky but seems to
work well and is better than the other alternatives
that I'd been able to come up with (like post-processing
the dependence files to munge the path for chapel_code.h).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14531 3a8e244f-b0f2-452b-bcba-4c88e055c3ca