Updating the nightly scripts to support xmt-sim testing
Updated the nightly-mta.cron script so that it would test the
xmt nightly compiles using the simulator rather than simply
making sure the compilations completed correctly.  This is
now enabled due to our support for a cross-compiled launch
binary.

Updated the nightly script so that if we're compiling for
xmt-sim, it will use the --savec output option.  This
causes the program library to stick around which gets rid
of warnings that zebra emits otherwise and permits the
simulator output to match the .good for files the tests
that work.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14532 3a8e244f-b0f2-452b-bcba-4c88e055c3ca