Fixed another dependence problem with chapel_code.c; fixed subdir makes
The other dependence problem that has been lurking for
chapel_code.c is that its dependences were not correct --
it named the (shallow, incorrect) file stdchpl[rt].h, but
did not name the files #included by that file and so on.
Fixed this using another depend command similar to what
we do by the parser which generates the dependences
manually, and did this by creating a stub chapel_code.c
file much like the chapel_code.h file that I created in
my previous checkin related to dependences.

I had also set up my conditionals within the runtime
subdirectory Makfiles backwards so that a "make"
command within the subdirectory would try to build
anything that the launcher required.  The common case
is the opposite, so reversed the sense of these
conditionals.  If you want to build a launcher's
files within the src/ or comm-*/ subdirectories, you
have to do "make MAKE_LAUNCHER=1".



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14533 3a8e244f-b0f2-452b-bcba-4c88e055c3ca