Replaced inout argument with a local array
I had previously used an inout argument in fft.chpl in order
to suggest that the (potentially) distributed array elements
should be copied into a local array in the function.  We
haven't really made clear what our argument-passing array
semantics do to distributed arrays/slices, though, so this
felt a bit iffy, and also subtle.  It also relied on reindexing
the array.

So, instead, wrote this to use a local array variable and
explicitly copy into/out of it.  This adds a few lines,
but should be easier to explain to the user, and a bit
easier to deal with in our memory management scheme (or
at least a more important and common case to work on for
now).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14536 3a8e244f-b0f2-452b-bcba-4c88e055c3ca