Cleaning up implementation of error message for arith domain assign error
Here I'm cleaning up the implementation of the error
message that I implemented the other day related to
assigning an iteratable object to an arithmetic domain,
which is illegal.  At that point, I implemented it by
putting a compilerError() message in the arithmetic
domain class' clear() method.  As Steve and I realized
shortly thereafter, this was a little bogus since one
might want a specific message relating to clear() for
arithmetic domains.  So here I added a new method,
clearForIteratableAssign() that is used by the assignment
in question -- for classes that can't be assigned in
this way, we generate a compilerError; for others, this
method calls clear().

This turned up some interesting properties of our
function resolution scheme -- in particular, one might
imagine that the default implementation of 
clearForIteratableAssign() could be placed in the
distribution hierarchy's root class and only overridden
for dense arithmetic and enumerated domains.  However,
it turns out that function resolution for a 'middle' 
class B:

    A
   / \
  B   D
 /    
C

not only considers classes C (necessary for dynamic
dispatch), B (necessary if the static class is
correct), and A (necessary if B doesn't define the
method in question), but it also can consider D.
This can happen for example if B does not implement
the method -- we correctly determine that A's copy
of the method may be called, but then lose track
of or ignore the information that the method came
from B and look for all possible calls that A.foo()
could resolve to, which includes D.foo().  This
is problematic if D.foo() contains a compilerError().
Thus, rather than defining foo() in A, we need to
push it down to the subtrees that don't lead us to
D.  In this case, that means implementing the
clearForIteratableAssign() and clear() methods in
classes adjacent to or below the arithmetic/enumerated
cases.

To help with this, I introduced abstract classes for
each of our (current) domain classes, resulting
in the following class hierarchy:

BaseDomain
  BaseArithmeticDomain
    BaseDenseArithmeticDomain
      SingleLocaleArithmeticDomain
    BaseSparseArithmeticDomain
      SingleLocaleSparseDomain
      CSRDomain
  BaseAssociativeDomain
    SingleLocaleAssociativeDomain
  BaseOpaqueDomain
    SingleLocaleOpaqueDomain
  BaseEnumDomain
    SingleLocaleEnumDomain


For now I implemented clear() on arithmetic domains
as a compilerError.  Arguably, it could also redefine
the domain to be [0..-1, 0..-1, ...] which could be
handy.

The one remaining mystery is that I had to
implement clearForIteratableAssign() in the
SingleLocaleAssociativeDomain() class in order
to get 
studies/hpcc/STREAMS/bradc/stream-block1dclass.chpl
to work.  I looked into this enough to see that the
problem (an internal error from generics.cpp) is
related to creating an associative domain of locales,
which the compiler is considering to be a generic
type, but I'm not sure why this would be.  Wil try
to capture in a smaller test case in hopes of
revealing a bug and removing this unnecessary case.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14542 3a8e244f-b0f2-452b-bcba-4c88e055c3ca