Fixing launcher Makefile dependences
Due to a very subtle thing being printed when compiling
launcher runtime directories (the gen/ directory's
timestamp was being removed after a make or make depend),
found that I hadn't gotten the make depend step correct
for the launcher directories.  Fixed this by refactoring
a number of aspects of the make depend step -- directory,
flags, includes, etc. into new makefile variables that
are set depending in one of two ways depending on whether
we're building the runtime or the launcher.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14543 3a8e244f-b0f2-452b-bcba-4c88e055c3ca