Adding one more (working) variation of testClear.chpl
This is a variation of testClear.chpl that Steve
pointed out would work due to the lack of a
clear() function in the base class (the common
parent between the classes for which clear() is
and is not supported).  This is what led me back
to wrestle some more with my class hierarchy and
come up with something that was reasonable clean
and worked.

Still, would like for testClear.chpl to work
someday...  :)




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14544 3a8e244f-b0f2-452b-bcba-4c88e055c3ca