Test cases to demonstrate a generic instantiation mystery
***NOTE: SingleLocaleAssociative.chpl must be modified to
remove clearForIteratableAssign() in order to reproduce
this.***

Yesterday I mentioned that I had to add what seemed like
an unnecessary clearForIteratableAssign() function to
the SingleLocaleAssociative class in order to keep one
test in the test system working.  This series of three
tests are simplified versions of the test -- the first
two work and the third one demonstrates what I was
seeing.

See the .future file for further notes.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14546 3a8e244f-b0f2-452b-bcba-4c88e055c3ca