Changed handling of on-, begin-, cobegin-, and coforall-blocks so that they are not changed into nested functions until just before nested functions are flattened (in passes/flattenFunctions).  Until that point, they are stored as blocks with BlockStmt::blockInfo (previously BlockStmt::loopInfo) set to the corresponding PRIMITIVE_BLOCK primitive.  This greatly simplifies the code in build.cpp.  The rationale for this change is to pave the way to allowing some iterators (leader/parallel) to put yield statements into on-, cobegin-, and coforall-blocks.
As a side note, after discussions with Brad, it seems that yields
should not be allowed in begin-statements because then it would not be
guaranteed that all the iterations of a forall loop would finish by
the time we exit the forall loop.  This restriction could be lightened
in the future but even so it seems like some restrictions would be
desireable to keep the forall loop well-behaved.  For example,
begin-blocks can be used but they need to be in sync-blocks.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14547 3a8e244f-b0f2-452b-bcba-4c88e055c3ca