First draft of PBS launcher for XT3/4
This is an initial/rough draft of a launcher for PBS on the XT
systems.  This launcher is a little more interesting than the
others because PBS works by submitting a script file to a
queueing system and having the script's commands run when
the processors become available.

Based on the sample script that David sent me for the XT
some time ago, I made the creation of the launch command
create the script file and then return the command
"qsub <scriptfile>" as the command that should be executed
in order to run the _real program.

The queueing system causes the stdout/stderr of the
command to be piped into a file, so I added a second
"cleanup" step to the launch interface.  For the other
launch commands, this is a no-op call.  For PBS, it
'cat's the output file to the console, then removes
the temporary files used for the output and for the
script itself.

The main hacky thing that this currently contains is
that there seemed to be a race after the qsub command
completes -- the stdout/stderr file may not yet have
been copied over at that point, so the cat and rm
commands can fail even though the console file will
be there when you go look for it manually.  I put
a sleep(1) in before the cat command currently to
try and avoid this race, but this is obviously not
a very good solution.  Better would be to watch for
the file to exist (or to be completely written?)
and cat it at that point, but I wanted to get an
initial draft in before wrestling with the sh command
that would be necessary to make this work.

Other things that could/should be improved:
- a lot of the filenames that I use are fairly
  hard-coded into the chpl_cfg_launch.c file.
  We should try to do a better job of uniquifying
  these using mktemp or the like.

- the system() commands should really be hoisted
  into a mysystem() command shared between all
  the runtime, much like we do with the compiler
  (and the -v flag should cause these commands to
  be reflected to the console).

- the strings that I use in these launch commands
  are not set up as cleanly w.r.t. buffer overflow
  as they should be.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14550 3a8e244f-b0f2-452b-bcba-4c88e055c3ca