One more XT launcher: aprun
This is one more launcher for the XT that wraps a 
command line command called aprun which is somewhat
similar to mpirun (as a result, I was able to get
this implemented in the time since my last commit).
This is the simpler way to launch XT programs, but
I didn't do it first because I couldn't find an
XT machine today that was configured to allow
aprun.

I made aprun the default launcher chosen by the
launcher.pl script when the target platform is
xt-cnl because I think it's lighter-weight and
more interactive, for systems that support it.
The PBS option is meant to be heavier-weight
and more for long-running/big #'s of nodes jobs.

Which reminds me, I forgot one more todo in my
PBS check-in.  There is an interactive mode that
PBS can be run in, and I'm wondering if there
is some sneaky way that we could have the launch
command invoke PBS in the interactive mode and
then run the aprun command on the user's behalf.
I was able to do the first step, but not the
second, so wanted to do a checkpoint of my 
almost-working code before wrestling with it
further.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14551 3a8e244f-b0f2-452b-bcba-4c88e055c3ca