Made it so the compiler no longer inserts an implicit coforall around the leader loop generated for a forall-statement.  Instead, the leader iterator is inlined and should explicitly contain the parallelism. See examples in test/functions/deitz/iterators/leader_follower/.
The remaining iterator work involves adding support for a parallel
iterator implementation of a forall-statement that is used when there
is no simultaneous iteration and changing the interface that the
author of leader-follower iterators needs to follow.

Disabled checking to look for yield statements inside cobegin-,
coforall-, forall-, and on-statements.

Improved expandIteratorInline routine to deal with iterators that
contain multiple yield statements.  This routine inlines an iterator
around a loop body.

To do: recursive leader iterators may lead to infinite compilation and
errors may not be effective if a yield is used in a cobegin-,
coforall-, forall-, or on-statement when not in a leader iterator.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14557 3a8e244f-b0f2-452b-bcba-4c88e055c3ca