Strengthened support for parsing single-dash compiler flags
Our previous support for parsing single-dash compiler flags
was somewhat weak, permitting illegal flags to slip through
in several cases.  Strengthened this, removing the rule
that if the compiler supports flags -a, -b, and -c, it
should support -abc as a shorthand for it (following gcc's
lead as my rationale).  Added a bunch of tests to test
against interesting cases that I ran into.

Strengthened the parsing of the -d (--log) flag to only
allow for legal log types -- previously, any character
was treated as legal.  This makes -devel an illegal
flag combination whereas previously it would've been
accepted and resulted in a log/ directory being created.
Removed four of the five log types as part of this change,
as they have been deprecated but their log types had not
been removed.

Improved the testing system to support a test-by-test
.lastcompopts file.  The .lastcompopts file has the
effect of specifying the final compiler options on a
line and is necessary for trying to test the -o flag
or a flag that is missing its argument.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14560 3a8e244f-b0f2-452b-bcba-4c88e055c3ca