Added destructors for some classes instantiated when arrays are created
With these destructors, the amount of memory leaked after destroying the local
array in function foo() in test/classes/figueroa/LeakedMemory2.chpl is just 24
bytes (instead of 236, which happens when no attempt is made to free any
memory), of which 20 are due to one sync var not being freed (the sync_var class
does not inherit from the object class); the other 4 are due to an instance
of the SingleLocaleDistribution class being created in _build_domain(), even
though it is never used.

The primitive PRIMITIVE_SYNC_DESTROY was added to allow the mutex and cond_vars
in the pthread implementation of sync vars to be freed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14561 3a8e244f-b0f2-452b-bcba-4c88e055c3ca