Re-enable fatal errors for getField()
I was getting seg faults due to uses of getField() for
fields that did not exist on distribution classes that
were expected to have specific fields (like eltType),
and with a bit of detective work realized that getField()
used to generate fatal errors for missing fields until
r13182 where David mentions:

"Also, previously it was an internal fatal error to call 
getField on a name not defined by the class. Now it returns 
NULL in that case. It might be good to add an argument to 
this method to decide how severe not finding the named 
field is that defaults to fatal error, but I haven't done 
that."

This was the same intuition I had about getField(), so I
added a fatal bool flag that is set to true by default
and then made an explicit call with "false" in the places
that seemed to require it (those more familiar with the
modified files below may want to check my work).  Also
removed a few INT_ASSERT()s that are no longer necessary
due to the change within getField().

Unrelated but too minor to check in on their own, updated 
svn:ignore for some of Andy's test directories.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14564 3a8e244f-b0f2-452b-bcba-4c88e055c3ca